+++ lib/multiselect_dropdown.dart
@@ -63,7 +63,9 @@ class MultiSelectDropDown extends StatefulWidget {
   final double? borderRadius;
   final BorderRadiusGeometry? radiusGeometry;
   final Color? borderColor;
+  final Color? focusedBorderColor;
   final double? borderWidth;
+  final double? focusedBorderWidth;
   final EdgeInsets? padding;
   final bool showClearIcon;
 
@@ -71,6 +73,13 @@ class MultiSelectDropDown extends StatefulWidget {
   final NetworkConfig? networkConfig;
   final Future<List<ValueItem>> Function(dynamic)? responseParser;
   final Widget Function(BuildContext, dynamic)? responseErrorBuilder;
+
+  /// focus node
+  final FocusNode? focusNode;
+
+  /// Controller for the dropdown
+  /// [controller] is the controller for the dropdown. It can be used to programmatically open and close the dropdown.
+  final MultiSelectController? controller;
 
   /// MultiSelectDropDown is a widget that allows the user to select multiple options from a list of options. It is a dropdown that allows the user to select multiple options.
   ///
@@ -199,15 +208,16 @@ class MultiSelectDropDown extends StatefulWidget {
       this.optionSeparator,
       this.inputDecoration,
       this.hintStyle,
-      this.padding = const EdgeInsets.symmetric(
-        horizontal: 8,
-        vertical: 8,
-      ),
+      this.padding,
+      this.focusedBorderColor = Colors.black54,
       this.borderColor = Colors.grey,
       this.borderWidth = 0.4,
+      this.focusedBorderWidth = 0.4,
       this.borderRadius = 12.0,
       this.radiusGeometry,
-      this.showClearIcon = true})
+      this.showClearIcon = true,
+      this.focusNode,
+      this.controller})
       : networkConfig = null,
         responseParser = null,
         responseErrorBuilder = null,
@@ -218,44 +228,45 @@ class MultiSelectDropDown extends StatefulWidget {
   /// [responseParser] is the parser that is used to parse the response from the network call.
   /// [responseErrorBuilder] is the builder that is used to build the error widget when the network call fails.
 
-  const MultiSelectDropDown.network(
-      {Key? key,
-      required this.networkConfig,
-      required this.responseParser,
-      this.responseErrorBuilder,
-      required this.onOptionSelected,
-      this.selectedOptionTextColor,
-      this.optionSeperator,
-      this.chipConfig = const ChipConfig(),
-      this.selectionType = SelectionType.multi,
-      this.hint = 'Select',
-      this.hintColor = Colors.grey,
-      this.hintFontSize = 14.0,
-      this.selectedOptions = const [],
-      this.disabledOptions = const [],
-      this.alwaysShowOptionIcon = false,
-      this.optionTextStyle,
-      this.selectedOptionIcon = const Icon(Icons.check),
-      this.selectedOptionBackgroundColor,
-      this.optionsBackgroundColor,
-      this.backgroundColor = Colors.white,
-      this.dropdownHeight = 200,
-      this.showChipInSingleSelectMode = false,
-      this.suffixIcon = Icons.arrow_drop_down,
-      this.selectedItemBuilder,
-      this.optionSeparator,
-      this.inputDecoration,
-      this.hintStyle,
-      this.padding = const EdgeInsets.symmetric(
-        horizontal: 8,
-        vertical: 8,
-      ),
-      this.borderColor = Colors.grey,
-      this.borderWidth = 0.4,
-      this.borderRadius = 12.0,
-      this.radiusGeometry,
-      this.showClearIcon = true})
-      : options = const [],
+  const MultiSelectDropDown.network({
+    Key? key,
+    required this.networkConfig,
+    required this.responseParser,
+    this.responseErrorBuilder,
+    required this.onOptionSelected,
+    this.selectedOptionTextColor,
+    this.optionSeperator,
+    this.chipConfig = const ChipConfig(),
+    this.selectionType = SelectionType.multi,
+    this.hint = 'Select',
+    this.hintColor = Colors.grey,
+    this.hintFontSize = 14.0,
+    this.selectedOptions = const [],
+    this.disabledOptions = const [],
+    this.alwaysShowOptionIcon = false,
+    this.optionTextStyle,
+    this.selectedOptionIcon = const Icon(Icons.check),
+    this.selectedOptionBackgroundColor,
+    this.optionsBackgroundColor,
+    this.backgroundColor = Colors.white,
+    this.dropdownHeight = 200,
+    this.showChipInSingleSelectMode = false,
+    this.suffixIcon = Icons.arrow_drop_down,
+    this.selectedItemBuilder,
+    this.optionSeparator,
+    this.inputDecoration,
+    this.hintStyle,
+    this.padding,
+    this.borderColor = Colors.grey,
+    this.focusedBorderColor = Colors.black54,
+    this.borderWidth = 0.4,
+    this.focusedBorderWidth = 0.4,
+    this.borderRadius = 12.0,
+    this.radiusGeometry,
+    this.showClearIcon = true,
+    this.focusNode,
+    this.controller,
+  })  : options = const [],
         super(key: key);
 
   @override
@@ -277,11 +288,14 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
   OverlayEntry? _overlayEntry;
   bool _selectionMode = false;
 
-  final FocusNode _focusNode = FocusNode();
+  late final FocusNode _focusNode;
   final LayerLink _layerLink = LayerLink();
 
   /// Response from the network call.
   dynamic _reponseBody;
+
+  /// value notifier that is used for controller.
+  MultiSelectController? _controller;
 
   @override
   void initState() {
@@ -289,6 +303,10 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
     WidgetsBinding.instance.addPostFrameCallback((_) {
       _initialize();
     });
+    _focusNode = widget.focusNode ?? FocusNode();
+    if (widget.controller != null) {
+      _controller = widget.controller;
+    }
   }
 
   /// Initializes the options, selected options and disabled options.
@@ -308,23 +326,39 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
 
   /// Adds the selected options and disabled options to the options list.
   void _addOptions() {
-    _selectedOptions.addAll(widget.selectedOptions);
-    _disabledOptions.addAll(widget.disabledOptions);
+    setState(() {
+      _selectedOptions.addAll(widget.selectedOptions);
+      _disabledOptions.addAll(widget.disabledOptions);
+    });
+
+    if (_controller != null) {
+      _controller!.setOptions(_options);
+      _controller!.setSelectedOptions(_selectedOptions);
+      _controller!.setDisabledOptions(_disabledOptions);
+
+      _controller!.addListener(_handleControllerChange);
+    }
   }
 
   /// Handles the focus change to show/hide the dropdown.
   _handleFocusChange() {
-    if (_focusNode.hasFocus) {
+    if (_focusNode.hasFocus && mounted) {
       _overlayEntry = _reponseBody != null && widget.networkConfig != null
           ? _buildNetworkErrorOverlayEntry()
           : _buildOverlayEntry();
       Overlay.of(context).insert(_overlayEntry!);
       return;
     }
-    _overlayEntry?.remove();
-    setState(() {
-      _selectionMode = _focusNode.hasFocus;
-    });
+    if (_overlayEntry != null) _overlayEntry?.remove();
+    if (mounted) {
+      setState(() {
+        _selectionMode = _focusNode.hasFocus;
+      });
+    }
+
+    if (_controller != null) {
+      _controller!.value._isDropdownOpen = _focusNode.hasFocus;
+    }
   }
 
   /// Handles the widget rebuild when the options are changed externally.
@@ -332,19 +366,37 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
   void didUpdateWidget(covariant MultiSelectDropDown oldWidget) {
     super.didUpdateWidget(oldWidget);
 
+    // If the options are changed externally, then the options are updated.
     if (widget.options != oldWidget.options) {
       _options.clear();
       _options.addAll(widget.options);
+
+      // If the controller is not null, then the options are updated in the controller.
+      if (_controller != null) {
+        _controller!.setOptions(_options);
+      }
     }
 
+    // If the selected options are changed externally, then the selected options are updated.
     if (widget.selectedOptions != oldWidget.selectedOptions) {
       _selectedOptions.clear();
       _selectedOptions.addAll(widget.selectedOptions);
+
+      // If the controller is not null, then the selected options are updated in the controller.
+      if (_controller != null) {
+        _controller!.setSelectedOptions(_selectedOptions);
+      }
     }
 
+    // If the disabled options are changed externally, then the disabled options are updated.
     if (widget.disabledOptions != oldWidget.disabledOptions) {
       _disabledOptions.clear();
       _disabledOptions.addAll(widget.disabledOptions);
+
+      // If the controller is not null, then the disabled options are updated in the controller.
+      if (_controller != null) {
+        _controller!.setDisabledOptions(_disabledOptions);
+      }
     }
   }
 
@@ -380,7 +432,7 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
               minWidth: MediaQuery.of(context).size.width,
               minHeight: 52,
             ),
-            padding: widget.padding,
+            padding: _getContainerPadding(),
             decoration: _getContainerDecoration(),
             child: Row(
               crossAxisAlignment: CrossAxisAlignment.center,
@@ -390,7 +442,7 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
                 ),
                 widget.showClearIcon && _anyItemSelected
                     ? InkWell(
-                        onTap: () => _clearSelection(),
+                        onTap: () => clear(),
                         child: const Icon(
                           Icons.close_outlined,
                           size: 14,
@@ -428,9 +480,8 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
     return _buildSelectedItems();
   }
 
-  /// check if single item is selected
-  /// return true if single item is selected and single select mode is enabled
-  bool get _anyItemSelected => _selectedOptions.length == 1;
+  /// return true if any item is selected.
+  bool get _anyItemSelected => _selectedOptions.isNotEmpty;
 
   /// Container decoration for the dropdown.
   Decoration _getContainerDecoration() {
@@ -439,10 +490,15 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
           color: widget.backgroundColor ?? Colors.white,
           borderRadius: widget.radiusGeometry ??
               BorderRadius.circular(widget.borderRadius ?? 12.0),
-          border: Border.all(
-            color: widget.borderColor ?? Colors.grey,
-            width: widget.borderWidth ?? 0.4,
-          ),
+          border: _selectionMode
+              ? Border.all(
+                  color: widget.focusedBorderColor ?? Colors.grey,
+                  width: widget.focusedBorderWidth ?? 0.4,
+                )
+              : Border.all(
+                  color: widget.borderColor ?? Colors.grey,
+                  width: widget.borderWidth ?? 0.4,
+                ),
         );
   }
 
@@ -450,11 +506,16 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
   @override
   void dispose() {
     if (_overlayEntry?.mounted == true) {
-      _overlayEntry?.remove();
-      _overlayEntry?.dispose();
+      if (_overlayState != null && _overlayEntry != null) {
+        _overlayEntry?.remove();
+      }
+      _overlayEntry = null;
       _overlayState?.dispose();
     }
     _focusNode.dispose();
+    if (_controller != null) {
+      _controller!.dispose();
+    }
     super.dispose();
   }
 
@@ -521,10 +582,10 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
       item: item,
       chipConfig: chipConfig,
       onItemDelete: (removedItem) {
-        setState(() {
-          _selectedOptions.remove(removedItem);
-        });
-        widget.onOptionSelected?.call(_selectedOptions);
+        if (_controller != null) {
+          _controller!.clearSelection(removedItem);
+        }
+
         if (_focusNode.hasFocus) _focusNode.unfocus();
       },
     );
@@ -563,21 +624,30 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
 
   /// Create the overlay entry for the dropdown.
   OverlayEntry _buildOverlayEntry() {
+    // Calculate the offset and the size of the dropdown button
     final values = _calculateOffsetSize();
+    // Get the size from the first item in the values list
     final size = values[0] as Size;
+    // Get the showOnTop value from the second item in the values list
     final showOnTop = values[1] as bool;
+
+    // Get the visual density of the theme
     final visualDensity = Theme.of(context).visualDensity;
+
+    // Calculate the height of the tile
     final tileHeight = 48.0 + visualDensity.vertical;
+    // Calculate the current height of the dropdown button
     final currentHeight = tileHeight * _options.length;
 
+    // Check if the dropdown height is less than the current height and greater than 0
     final bool isScrollable =
         widget.dropdownHeight < currentHeight && widget.dropdownHeight > 0;
+    // Calculate the offset in the Y direction
     final offsetY = showOnTop
         ? isScrollable
             ? -widget.dropdownHeight - 5
             : -currentHeight - 5
         : size.height + 5;
-
     return OverlayEntry(builder: (context) {
       List<ValueItem> options = _options;
       List<ValueItem> selectedOptions = [..._selectedOptions];
@@ -621,9 +691,11 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
                                       fontSize: widget.hintFontSize,
                                     )),
                             textColor: Colors.black,
+                            focusColor: Colors.red,
                             selectedColor:
                                 widget.selectedOptionTextColor ?? primaryColor,
                             selected: isSelected,
+                            autofocus: true,
                             dense: true,
                             tileColor:
                                 widget.optionsBackgroundColor ?? Colors.white,
@@ -658,6 +730,12 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
                                   _selectedOptions.add(option);
                                 });
                                 _focusNode.unfocus();
+                              }
+
+                              if (_controller != null) {
+                                _controller!.value._selectedOptions.clear();
+                                _controller!.value._selectedOptions
+                                    .addAll(_selectedOptions);
                               }
 
                               widget.onOptionSelected?.call(_selectedOptions);
@@ -776,15 +854,196 @@ class _MultiSelectDropDownState extends State<MultiSelectDropDown> {
   }
 
   /// Clear the selected options.
-  /// If the dropdown is showing, the dropdown will be removed and a new one will be created.
-  void _clearSelection() {
-    _selectedOptions.clear();
-    widget.onOptionSelected?.call(_selectedOptions);
-    setState(() {});
-    if (_overlayEntry != null) {
-      _overlayEntry!.remove();
-      _overlayEntry = null;
-      _toggleFocus();
+  /// [MultiSelectController] is used to clear the selected options.
+  void clear() {
+    _controller?.setSelectedOptions([]);
+    if (_focusNode.hasFocus) _focusNode.unfocus();
+  }
+
+  /// handle the controller change.
+  void _handleControllerChange() {
+    // if the controller is null, return.
+    if (_controller == null) return;
+
+    // if current disabled options are not equal to the controller's disabled options, update the state.
+    if (_disabledOptions != _controller!.value._disabledOptions) {
+      setState(() {
+        _disabledOptions.clear();
+        _disabledOptions.addAll(_controller!.value._disabledOptions);
+      });
     }
+
+    // if current options are not equal to the controller's options, update the state.
+    if (_options != _controller!.value._options) {
+      setState(() {
+        _options.clear();
+        _options.addAll(_controller!.value._options);
+      });
+    }
+
+    // if current selected options are not equal to the controller's selected options, update the state.
+    if (_selectedOptions != _controller!.value._selectedOptions) {
+      setState(() {
+        _selectedOptions.clear();
+        _selectedOptions.addAll(_controller!.value._selectedOptions);
+      });
+      widget.onOptionSelected?.call(_selectedOptions);
+    }
+
+    if (_selectionMode != _controller!.value._isDropdownOpen) {
+      if (_controller!.value._isDropdownOpen) {
+        _focusNode.requestFocus();
+      } else {
+        _focusNode.unfocus();
+      }
+    }
+  }
+
+  // get the container padding.
+  EdgeInsetsGeometry _getContainerPadding() {
+    if (widget.padding != null) {
+      return widget.padding!;
+    }
+    return widget.selectionType == SelectionType.single
+        ? const EdgeInsets.symmetric(horizontal: 8.0, vertical: 8.0)
+        : const EdgeInsets.symmetric(horizontal: 8.0, vertical: 0);
+  }
+}
+
+/// MultiSelect Controller class.
+/// This class is used to control the state of the MultiSelectDropdown widget.
+/// This is just abstract class. The implementation of this class is in the MultiSelectDropdown class.
+/// The implementation of this class is hidden from the user.
+/// The implmentation of this class is in the MultiSelectDropdown class.
+class _MultiSelectController {
+  final List<ValueItem> _disabledOptions = [];
+  final List<ValueItem> _options = [];
+  final List<ValueItem> _selectedOptions = [];
+  bool _isDropdownOpen = false;
+}
+
+/// implementation of the MultiSelectController class.
+class MultiSelectController extends ValueNotifier<_MultiSelectController> {
+  MultiSelectController() : super(_MultiSelectController());
+
+  /// Clear the selected options.
+  /// [MultiSelectController] is used to clear the selected options.
+  void clearAllSelection() {
+    value._selectedOptions.clear();
+    notifyListeners();
+  }
+
+  /// clear specific selected option
+  /// [MultiSelectController] is used to clear specific selected option.
+  void clearSelection(ValueItem option) {
+    if (!value._selectedOptions.contains(option)) return;
+
+    if (value._disabledOptions.contains(option)) {
+      throw Exception('Cannot clear selection of a disabled option');
+    }
+
+    if (!value._options.contains(option)) {
+      throw Exception(
+          'Cannot clear selection of an option that is not in the options list');
+    }
+
+    value._selectedOptions.remove(option);
+    notifyListeners();
+  }
+
+  /// select the options
+  /// [MultiSelectController] is used to select the options.
+  void setSelectedOptions(List<ValueItem> options) {
+    if (options.any((element) => value._disabledOptions.contains(element))) {
+      throw Exception('Cannot select disabled options');
+    }
+
+    if (options.any((element) => !value._options.contains(element))) {
+      throw Exception('Cannot select options that are not in the options list');
+    }
+
+    value._selectedOptions.clear();
+    value._selectedOptions.addAll(options);
+    notifyListeners();
+  }
+
+  /// add selected option
+  /// [MultiSelectController] is used to add selected option.
+  void addSelectedOption(ValueItem option) {
+    if (value._disabledOptions.contains(option)) {
+      throw Exception('Cannot select disabled option');
+    }
+
+    if (!value._options.contains(option)) {
+      throw Exception('Cannot select option that is not in the options list');
+    }
+
+    value._selectedOptions.add(option);
+    notifyListeners();
+  }
+
+  /// set disabled options
+  /// [MultiSelectController] is used to set disabled options.
+  void setDisabledOptions(List<ValueItem> disabledOptions) {
+    if (disabledOptions.any((element) => !value._options.contains(element))) {
+      throw Exception(
+          'Cannot disable options that are not in the options list');
+    }
+
+    value._disabledOptions.clear();
+    value._disabledOptions.addAll(disabledOptions);
+    notifyListeners();
+  }
+
+  /// setDisabledOption method
+  /// [MultiSelectController] is used to set disabled option.
+  void setDisabledOption(ValueItem disabledOption) {
+    if (!value._options.contains(disabledOption)) {
+      throw Exception('Cannot disable option that is not in the options list');
+    }
+
+    value._disabledOptions.add(disabledOption);
+    notifyListeners();
+  }
+
+  /// set options
+  /// [MultiSelectController] is used to set options.
+  void setOptions(List<ValueItem> options) {
+    value._options.clear();
+    value._options.addAll(options);
+    notifyListeners();
+  }
+
+  /// get disabled options
+  List<ValueItem> get disabledOptions => value._disabledOptions;
+
+  /// get enabled options
+  List<ValueItem> get enabledOptions => value._options
+      .where((element) => !value._disabledOptions.contains(element))
+      .toList();
+
+  /// get options
+  List<ValueItem> get options => value._options;
+
+  /// get selected options
+  List<ValueItem> get selectedOptions => value._selectedOptions;
+
+  /// get is dropdown open
+  bool get isDropdownOpen => value._isDropdownOpen;
+
+  /// show dropdown
+  /// [MultiSelectController] is used to show dropdown.
+  void showDropdown() {
+    if (value._isDropdownOpen) return;
+    value._isDropdownOpen = true;
+    notifyListeners();
+  }
+
+  /// hide dropdown
+  /// [MultiSelectController] is used to hide dropdown.
+  void hideDropdown() {
+    if (!value._isDropdownOpen) return;
+    value._isDropdownOpen = false;
+    notifyListeners();
   }
 }